feat: add a collie info command
This command now returns runtime and detected CLI version infos.
The version command now outputs only the version number
