+++ src/main.ts
@@ -16,6 +16,7 @@ import { registerVersionCommand } from "./commands/version.command.ts";
 import { FirstTimeExperience } from "./FirstTimeExperience.ts";
 import { CollieFoundationDoesNotExistError } from "./model/schemas/ModelValidator.ts";
 import { makeTopLevelCommand } from "./commands/TopLevelCommand.ts";
+import { registerInfoCommand } from "./commands/info.command.ts";
 
 async function collie() {
   const program = makeTopLevelCommand()
@@ -35,6 +36,7 @@ async function collie() {
   registerKitCommand(program);
   registerComplianceCommand(program);
 
+  registerInfoCommand(program);
   registerUpgradeCommand(program);
   registerVersionCommand(program);
